Initial commit of New Antics Game And a QLearning AI prototype
This version of antics differs from the version used in the CS421-AI
repository. It has different rules that change the game a lot.

This has an initial prototype of a QLearning AI.
As of right now, it runs and saves its memory every 100 games.
State compression is poor, and needs work.
There is currently a problem with the serialization in that the file gets too
large to serialize I believe. Pickle encountered at Memory problem
after trying to save its memory after 600 games.
This AI is NOT using Eligibility Tracing.